+++ lib/src/theme/stylix_theme_controller.dart
@@ -36,4 +36,3 @@ class StylixThemeController extends ChangeNotifier {
     notifyListeners();
   }
 }
-